done
added a many-to-many relation and new crud for the new entity

+++ rest-backend/src/main/java/com/example/restapi/Service/CustomerService.java
@@ -75,6 +75,8 @@ public class CustomerService {
     }
 
     public void deleteCustomer(Long id) {
+        if (this.repository.findById(id).isEmpty())
+            throw new CustomerNotFoundException(id);
         repository.deleteById(id);
     }
